backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
cloudinary==1.36.0
mangum==0.17.0
pydantic==2.10.3
email-validator==2.1.0
python-dotenv==1.0.0